feat!: Java 17, Spring 3.1x, Gradle upgrade working BREAKING_CHANGE: Requires Java 17

+++ construct/obcorelib/src/main/java/com/onbelay/core/entity/repository/BaseRepository.java
@@ -854,7 +854,7 @@ public class BaseRepository<T> implements EntityRepository<T> {
         List items = query.getResultList();
         if ( items.isEmpty())
             return null;
-        BigInteger bg = (BigInteger) items.get(0);
+        Long bg = (Long) items.get(0);
         return bg.intValue();
     }
